FAQs page fixed #176 - Admin input from the web is possible ow - Improved html and css - Navbar working - database.py code dedicated to faqs moved to load_faqs.py - Deleted duplicated api file
