+++ Alejandria/static/css/form-sell.css
@@ -197,4 +197,3 @@ border: 1px solid rgba(0, 0, 0, .2);
 	 background: #f7b1ab;
 	 border: 1px solid rgba(0, 0, 0, .1);
 }
- }
