+++ Alejandria/templates/user_library.html
@@ -8,63 +8,67 @@
 <html lang="en">
 
 <link rel="stylesheet" href="{% static 'css/editor-page/card.css' %}">
-<head>
-    <meta charset="UTF-8">
-    <title>Alejandria</title>
-</head>
-<body>
-    <div class="container-fluid">
-        <div class="row" >
-            <div class="col-12 mt-3"  >
-                {% for book in user_prod %}
-                    <div class="card" style="background-color: #963434; margin-left: 15%; margin-right:15%; margin-top:2%;">
-                         <div class="card-horizontal" >
-                             {% if book.thumbnail %}
-                                <img src="{{ book.thumbnail.url }}" alt="Card image cap" style="width:180px; height:250px;">
-                                {% else %}
-                                <img src= "{% static 'no_available.jpg' %}" alt="Card image cap" style="width:180px; height:250px;">
-                            {% endif %}
-                             <div class="card-body" style="color:white;" >
-                                 <h3 class="card-title">{{ book.title }}</h3>
-                                 <div class="row" style="margin-top:2%;">
-                                     <div class="column" style="margin-left:5%;" >
-                                         <h6 class="card-title"><b>Author:</b> {{ book.author }} </h6>
-                                         <h6 class="card-title"><b>Saga:</b> {{ book.saga }} </h6>
-                                         <h6 class="card-title"><b>Price:</b> {{ book.price }} $</h6>
-                                     </div>
-                                     <div class="column" style="position: absolute; right: 30%">
-                                         <h6 class="card-title"><b>ISBN:</b> {{ book.ISBN }}</h6>
-                                         <h6 class="card-title"><b>Genre:</b> {{ book.get_primary_genre_display }}, {{ book.get_secondary_genre_display }}</h6>
-                                         <h6 class="card-title"><b>Language:</b> {{ book.language }} </h6>
-                                         <h6 class="card-title"><b>Publication date:</b> {{ book.publication_date }} </h6>
-                                     </div>
-                                     <div class="column" style="position: absolute; right: 5%; " >
-                                        <div class="row" style="margin-top: 5%;">
-                                            <form action="/book/{{book.ISBN}}">
-                                                <button class="btn btn-danger my-2 my-sm-0"  style="width:100%" title="Show"  type="submit"><i class="fas fa-search"></i></button>
-                                            </form>
-                                          </div>
-                                         <div class="row" style="margin-top:5%;">
-                                             <form action="/pdf"> <!--{{prod.ISBN.ISBN}}-->
-                                                <button class="btn btn-danger my-2 my-sm-0" style="width:100%" title="Download" type="submit"><i class="fas fa-download"></i></button>
-                                             </form>
-                                          </div>
-                                     </div>
-                                 </div>
 
-                            </div>
-                        </div>
-                    </div>
-                        {% endfor %}
-            </div>
-            {% if user_prod|length == 0 %}
-                <div class="col-12 mt-3" style="color: white; text-align: center;">
-                  <h3>Seems like you did not buy any book yet...</h3>
-                </div>
+<head>
+  <meta charset="UTF-8">
+  <title>Alejandria</title>
+</head>
+
+<body>
+  <div class="container-fluid">
+    <div class="row">
+      <div class="col-12 mt-3">
+        {% for book in user_prod %}
+        <div class="card" style="background-color: #963434; margin-left: 15%; margin-right:15%; margin-top:2%;">
+          <div class="card-horizontal">
+            {% if book.thumbnail %}
+            <img src="{{ book.thumbnail.url }}" alt="Card image cap" style="width:180px; height:250px;">
+            {% else %}
+            <img src="{% static 'no_available.jpg' %}" alt="Card image cap" style="width:180px; height:250px;">
             {% endif %}
+            <div class="card-body" style="color:white;">
+              <h3 class="card-title">{{ book.title }}</h3>
+              <div class="row" style="margin-top:2%;">
+                <div class="column" style="margin-left:5%;">
+                  <h6 class="card-title"><b>Author:</b> {{ book.author }} </h6>
+                  <h6 class="card-title"><b>Saga:</b> {{ book.saga }} </h6>
+                  <h6 class="card-title"><b>Price:</b> {{ book.price }} $</h6>
+                </div>
+                <div class="column" style="position: absolute; right: 30%">
+                  <h6 class="card-title"><b>ISBN:</b> {{ book.ISBN }}</h6>
+                  <h6 class="card-title"><b>Genre:</b> {{ book.get_primary_genre_display }}, {{ book.get_secondary_genre_display }}</h6>
+                  <h6 class="card-title"><b>Language:</b> {{ book.language }} </h6>
+                  <h6 class="card-title"><b>Publication date:</b> {{ book.publication_date }} </h6>
+                </div>
+                <div class="column" style="position: absolute; right: 5%; ">
+                  <div class="row" style="margin-top: 5%;">
+                    <form action="/book/{{book.ISBN}}">
+                      <button class="btn btn-danger my-2 my-sm-0" style="width:100%" title="Show" type="submit"><i class="fas fa-search"></i></button>
+                    </form>
+                  </div>
+                  <div class="row" style="margin-top:5%;">
+                    <form action="/pdf">
+                      <!--{{prod.ISBN.ISBN}}-->
+                      <button class="btn btn-danger my-2 my-sm-0" style="width:100%" title="Download" type="submit"><i class="fas fa-download"></i></button>
+                    </form>
+                  </div>
+                </div>
+              </div>
+
+            </div>
+          </div>
         </div>
+        {% endfor %}
+      </div>
+      {% if user_prod|length == 0 %}
+      <div class="col-12 mt-3" style="color: white; text-align: center;">
+        <h3>Seems like you did not buy any book yet...</h3>
+      </div>
+      {% endif %}
     </div>
+  </div>
 </body>
+
 </html>
