+++ books/urls.py
@@ -30,6 +30,9 @@ urlpatterns = [
     path('pdf/', views.generate_pdf, name='generate_bill'),
     path('library/', views.UserLibrary.as_view(), name='user_library'),
     path('bills/', views.UserBills.as_view(), name='user_bills'),
-    path('check/', views.check_data, name="check_username")
+    path('check/', views.check_data, name="check_username"),
+    path('addfaq/', views.addfaq, name="add_faq"),
+    path('modifyfaq/', views.modifyfaq, name="modify_faq"),
+    path('deletefaq/', views.deletefaq, name="delete_faq"),
 
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
